Added clock: The time it took for the program to execute will now be outputted
Changed matrix_index: short -> int, to prevent a short overflow

Tested NUM_USERS: Limit is around 345.000

+++ src/lib/simulation.c
@@ -252,7 +252,7 @@ void generate_matrices(WebsiteNode **linked_websites, const short NUM_CATEGORIES
             short num_y = NUM_USERS;
 
             short *matrix = (short*) malloc(sizeof(short) * num_x * num_y);
-            short matrix_index;
+            int matrix_index;
             for (int y = 0; y < num_y; y++) {
                 for (int x = 0; x < num_x; x++) {
                     matrix_index = x + y * num_x;
